Fix RegisterAltEvent

+++ internal/module/capi.go
@@ -8,5 +8,5 @@ func RegisterAltEvent(resource string, event uint16) {
 	cresource := C.CString(resource)
 	defer C.free(unsafe.Pointer(cresource))
 
-	GetFunc("RegisterAltEvent", uintptr(unsafe.Pointer(cresource)), C.uint(event))
+	GetFunc("RegisterAltEvent", uintptr(unsafe.Pointer(cresource)), uintptr(C.uint(event)))
 }
